Put the MCU in a low power mode when QP is idle.
We use LPM3 when the CPU is not required.  LPM4 cannot be used because
that shuts down the Basic Timer 1 clock, which is required to wake up
again.

The role of the LED has changed.  It now is an indication that the CPU
is awake.  We turn the LED off when we go into low power mode, and
turn it back on at the start of any interrupt handlers.

Signed-off-by: Russell Steicke <russells@adelie.cx>