Stop jumping quickly between temperatures.
The ADC's LSB can sometimes appear unstable, and "jitter" back and
forth by one LSB.  If it happens to jitter over a temperature
boundary, our temperature can jitter a whole degree each reading.  So
add some hysteresis in the temperature reading to prevent this.

A single LSB of hysteresis (so the ADC reading has to move by two LSB
for us to register a change) seems to be enough.

Signed-off-by: Russell Steicke <russells@adelie.cx>