Increase the temperature range and make calculations clearer.
The temperature range is now -30 to +70 (50 Colds to 50 Hots), so it
can be put essentially anywhere people will go in Australia.

Some of the temperature calculations were not very clear, mainly
to do with working out the index into the array of strings that
indicate the temperature.

To ensure we avoid overflow in the temperature calculations, all the
values from the table of ADC to temperature and onwards are now
int16_t instead of int8_t.

Signed-off-by: Russell Steicke <russells@adelie.cx>