Implement the calibration mode.
Calibration is done by holding down the first button.  The display
shows temperature in degrees C and the offset.  (The temperature is
shown as after the offset is applied.)  Simultaneously pressing a
second button increases the offset, and a third button decreases the
offset.  Releasing the first button returns to displaying hots and
colds.

The offset is stored in the first segment of the MSP430F449
information memory.

The first button sometimes takes a while to respond, as it is only
checked on every QP-nano tick, which is currently every two seconds.
The hardware does not currently support interrupts on pin change, but
the custom designed hardware will do so, so the button will react
immediately.

Signed-off-by: Russell Steicke <russells@adelie.cx>