Fix an error in timer periods.
There was an error in the code that masked the timer period selection
bits, that left bit 1 always high.  This meant we could never select a
Basic Timer 1 period less than fCLK2/8.

Also, leave in the code that raises P2.0 whenever the CPU is awake,
and lowers it when we enter low power mode.

Signed-off-by: Russell Steicke <russells@adelie.cx>