Measure temperature with a MCP9700A.
The MCP9700A is an external temperature device in a TO-92 package.  We
want this package so we measure air temperature, not PCB temperature.

The device is powered from a port pin (P6.1), with an RC power supply
filter as recommended in its datasheet.  It produces an analogue
output, which we sense on pin A0 (P6.0).  It produces zero degrees C
at 500mV, and a slope of 10mV per degree.  Its accuracy is typically
+/- 1C at 25C and 3.3V supply -- but note that and we'll probably be
using 3.0V.  The ADC reference will introduce extra inaccuracies so a
calibration mode is still required.  When we're not measuring
temperature, we shut down the device to save power.

This device has a greater output voltage range change per degree than
the internal MSP430F449 temperature sensor.  Therefore, we also
increase the hysteresis in the temperature sensing from two LSB to
four LSB.

Signed-off-by: Russell Steicke <russells@adelie.cx>