Change QP-nano to version 5.1.1.
QP-nano 5.1.1 has a feature that we need - multiple timer sources.  We
can have several timers with different rates, and they send different
signals.  So in the future we can have a stable clock for an RTC, and
a short period clock that is only turned on during temperature
measurement or when the UI is busy.

Signed-off-by: Russell Steicke <russells@adelie.cx>