Use two different timers so we don't reset the RTC timer.
Use a second timer when a fast timer tick is required, and leave Basic
Timer 1 alone.

Previously we have been using one timer (Basic Timer 1) to give us
both the slow timer ticks (when we are paused and not doing much) and
the fast timer ticks (when we are measuring temperature, or doing
calibration.)  Every time we change between clock rates, the Basic
Timer module is restarted.

But Basic Timer 1 will be used for the RTC, so we don't want to ever
restart it.

Signed-off-by: Russell Steicke <russells@adelie.cx>