Show a clock on the LCD.
This is a very basic RTC.  There is no way to set the time or to
adjust for clock drift.

We use the clock digits at the top right of the SBLCDA2 display (or
on the equivalent Olimex LCD-449STK2 display on the MSP430-449STK2
development board.)

We toggle the colon every two seconds to show life.

Signed-off-by: Russell Steicke <russells@adelie.cx>