Add a button state machine.
We send events from the fast timer ISR to a button state machine,
which decides how long buttons have been held down and whether to send
button events to the (minimal) UI.

The UI will be able to become more complicated with better button
handling in place.

The new button state machine does a fair bit of work on every tick,
but it only runs when the first button is held down.  The rest of the
time it's completely idle, so there is no real power cost.

Signed-off-by: Russell Steicke <russells@adelie.cx>