Make a greater separation between time keeping and UI.
Move the display of time into the UI, so that the RTC has to ask the
UI to display time for it.  This will make time setting easier.

There is a cost: every two seconds we do a trip through QP-nano to get
the signal and parameter into the UI.  Previously we could short
circuit this and flash the colon in the RTC, but now we have to get
the UI to do that.

Signed-off-by: Russell Steicke <russells@adelie.cx>