Make serial output faster.
Serial output at 9600 baud is a bit slow to keep up with some of the
information we need to get out.  So output at 115200 instead.

This means that with serial turned on (make SERIAL=yes) the 1MHz SMCLK
is always enabled, do don't leave serial turned on in a battery
powered system.

Signed-off-by: Russell Steicke <russells@adelie.cx>