Adjust the calibration with a proper menu item.
A press of button 1 (detected only every two seconds) moves the UI
into a menu state.  There are two menu entries "T CAL" for temperature
calibration, and "EXIT" to return to the normal thermometer mode.

There is a timeout bar display (currently simulated with the "units"
display on the LCD-449STK2) that counts down until the user interface
automatically returns to thermometer mode.  The timeout is reset any
time a button is pressed.

Other menu items can be added relatively easily from here.  The
biggest one to come is time setting.

We also make some changes to the buttons, as the previous button code
appeared too slow and was definitely over-engineered.  There is still
some element of that, and it will go away in the future.

Signed-off-by: Russell Steicke <russells@adelie.cx>