Maintain the date as well as the time.
The default starting date is 2014-03-07 (of course).  But with the
proposed hardware, the super capacitor should maintain power over
battery changes and the data should never need setting.

We maintain days of the month correctly, including leap years.  We
allow years from 2014 to 2399.  The upper limit of the year means that
we can assume that every fourth year is a leap year, and not worry
about the fourth century exceptions.

TODO: set the default starting time to the time of the ceremony.

Signed-off-by: Russell Steicke <russells@adelie.cx>