Allow time rate adjustment.
The time rate adjustment is in seconds/day, with a range of +/-9.  The
adjustment is subtracted from the seconds count at the rollover of
each day.  So an adjustment of +5 means that the first minute of each
day is five seconds shorter than normal.

To figure out the adjustment, set the time one day and find the
clock's offset from the real time at the same time the next day.  If
the clock loses time, the adjustment should be negative

This may not be fine enough for a real clock, as one second per day is
about six minutes per year.

Signed-off-by: Russell Steicke <russells@adelie.cx>