Create a recorder active object to maintain temperatures.
This moves the temperature measurement and recording out of the user
interface.

Previously this wasn't an issue as we only displayed the current
temperature.  But now that we intend keep a record of the day's
maximum and minimum temperatures, we want these maintained even while
the UI is otherwise busy.

Signed-off-by: Russell Steicke <russells@adelie.cx>