When returning from a UI action, show the temperature immediately.
Previously, when the user has finished an action in the user interface
and we return to the idle state that just shows the current
temperature, we had to wait for a temperature event from the recorder
before anything was displayed.  This seemed like the UI was hanging.

So fix that by showing the current temperature immediately if it's
available.

We show "???" if the temperature is not yet available.

Signed-off-by: Russell Steicke <russells@adelie.cx>