Use the buttons on the new hardware and run a faster BT1.
Also, correct a series of silly errors in the macros that make the
unused pins into low outputs.  (I had eg SB(P1DIR,4) instead of the
correct SB(P1DIR,BIT4).)

The new hardware has buttons on different pins to the dev board.

Also, we now run the Basic Timer 1 ISR at 32Hz instead of 0.5Hz.  See
the comment below about bounces.

NOTE: We are only using buttons 1, 2, and 3 now.  Button 4 (Cancel)
will be added when we have code to handle it.

The buttons are connected to:

B1 - P1.3
B2 - P1.2
B3 - P1.1
B4 - P1.0

Yes, they're in reverse order.  It was easier to route the signals
that way.

These pins are interrupt capable, but the buttons being used are very
bouncy, so the interrupts are not enabled now.  If there's time later
this can be debugged and added.

Signed-off-by: Russell Steicke <russells@adelie.cx>