Get the RTC to do time adjustments.
Move time adjustment calculations out of the user interface, and into
the real time clock.  This simplifies the UI, which only gets told
about time events and doesn't have to do any adjustments.

The user can adjust the RTC rate in 1/4 second per day increments,
between +9.75 and -9.75 seconds per day.  The actual adjustment is
done is 1/8 second steps, several times per day depending on the
amount.  The number of steps and the times that the steps happen are
generated at compile time.

The adjustment is the offset that the clock needs to have.

- If the clock runs fast, the adjustment needs to be negative.
- If the clock runs slow, the adjustment needs to be positive.

Signed-off-by: Russell Steicke <russells@adelie.cx>