Add the ability to automatically check the time adjustment.
When SERIAL=yes, the device will print the time every minute, in the
form

  \r\nTIME=HHMM\r\n

There is also a script that will read these times, work out the
current offset, and print the time and offset.  The script sorts
through the other debugging output and only reacts to those lines.

We can then leave the device running and the script running with its
stdin from the serial port, and check for clock drift by examining the
offsets.

Signed-off-by: Russell Steicke <russells@adelie.cx>