Avoid spurious button presses when entering or leaving the UI.
When entering or leaving the UI, we tell the buttons state machine to
wait for all buttons to be up before proceeding.  Because the buttons
are quite bouncy, we can see all buttons up immediately followed by a
button press possibly a couple of times.  So make the buttons UI wait
longer - a number of ISRs with no buttons down, in succession - before
proceeding.

When the fast timers are not running, the BSP has the same issue, so
get the timer ISR to also count ticks with all buttons up.

Signed-off-by: Russell Steicke <russells@adelie.cx>