Fix the temperature calibration menu mode.
Previously, we sent signals from the ADC ISR to only the recorder, and
never to any other AO.  So when the ui was in temperature calibration
mode, it never saw the replies.  Fix that.

This also includes a bit of code intended for the unimplemented
battery monitor - specifically the channel enum that selects the
desired ADC channel, and the VCC_CHANNEL value.

Signed-off-by: Russell Steicke <russells@adelie.cx>