Fix random ADC wait assertions.
These were caused by the UI exiting the calibration mode while it had
the ADC locked, and not releasing the ADC.

While the UI is in the states that have the ADC locked, we defer the
events that cause us to exit, and handle those when we go back to the
calibration mode's pause state.  We can actually only defer one event,
and if there are multiple events that get deferred, only the last is
remembered.  But that should be sufficient as the user will not be
able to tell the difference between

- "I pressed the button but I just missed the end of the menu timeout
  period", and

- "I pressed the button but the menu timeout period happened just
  after and overwrote my deferred button event."

Because this requires an extra state in the middle -- so it can ignore
and defer the right events -- we have had to increase the value of
QEP_MAX_NEST_DEPTH_ in qepn.c.

Signed-off-by: Russell Steicke <russells@adelie.cx>