On assertion, save info in memory for later display.
The assertion string and number (normally module and line number) are
saved in RAM, in a place that doesn't get initialised on startup.
They are displayed by a "secret" menu entry -- get to the bottom entry
("ADJTIME"), then press down seven more times to display the assertion
string on the main display, and the number on the time display.  The
number is limited to four digits, but since it's meant to be a line
number that shouldn't be a real limitation.

As an added secret, press down 37 more times while displaying the last
assertion to create a new one!

Signed-off-by: Russell Steicke <russells@adelie.cx>